Move mixins to own file
