+++ fiboa_cli/datasets/es_ga.py
@@ -1,4 +1,4 @@
-from fiboa_cli.convert_utils import EsriRESTConverterMixin
+from fiboa_cli.converter_rest import EsriRESTConverterMixin
 from fiboa_cli.datasets.es import ESBaseConverter
